tg-automatizamtion/requirements.txt
# Telegram Automation System - Python Dependencies

# Browser automation
playwright>=1.40.0

# Configuration
pyyaml>=6.0

# Note: SQLite is built into Python, no installation needed
